src changes in support of GarageDoorClass addition
